133 Introduction to Gates

+++ app/Http/Controllers/PostsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class PostsController extends Controller
 {
@@ -84,7 +85,13 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
+        $post = BlogPost::findOrFail($id);
+
+        if (Gate::denies('update-post', $post)) {
+            abort(403, "You can't edit this blog post!");
+        }
+
+        return view('posts.edit', ['post' => $post]);
     }
 
     /**
@@ -97,6 +104,11 @@ class PostsController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        if (Gate::denies('update-post', $post)) {
+            abort(403, "You can't edit this blog post!");
+        }
+
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
@@ -115,6 +127,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        if (Gate::denies('delete-post', $post)) {
+            abort(403, "You can't delete this blog post!");
+        }
+
         $post->delete();
 
         session()->flash('status', 'Blog post was deleted!');
